requirements.txt
fastapi[standard]==0.108.0
gunicorn==21.2.0
uvicorn==0.25.0
PyJWT==2.8.0
pytest==7.4.4